ok4

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: haghbal <haghbal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/02 23:14:27 by haghbal           #+#    #+#             */
-/*   Updated: 2023/11/22 00:13:19 by haghbal          ###   ########.fr       */
+/*   Updated: 2023/11/30 23:17:49 by haghbal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
